Rename transaction to payment
