+++ packages/api/src/modules/payment/__test__/transaction-service.spec.ts
@@ -7,8 +7,8 @@ import { EventRepositoryImpl } from "../../event/event-repository"
 import { Kysely } from "kysely"
 import { ProductRepositoryImpl } from "../product-repository"
 import Stripe from "stripe"
-import { TransactionRepositoryImpl } from "../transaction-repository"
-import { TransactionServiceImpl } from "../transaction-service"
+import { TransactionRepositoryImpl } from "../payment-repository"
+import { TransactionServiceImpl } from "../payment-service"
 import { eventPayload } from "../../event/__test__/event-service.spec"
 import { paymentProvidersPayload } from "./product-payment-provider.spec"
 import { productPayload } from "./product-service.spec"
