+++ packages/api/src/modules/payment/payment-router.ts
@@ -1,12 +1,29 @@
-import { protectedProcedure, t } from "../../trpc"
-
 import { productRouter } from "./product-router"
-import { transactionRouter } from "./transaction-router"
+import { protectedProcedure } from "../../trpc"
+import { t } from "../../trpc"
+import { z } from "zod"
 
 export const paymentRouter = t.router({
   product: productRouter,
-  transaction: transactionRouter,
   getPaymentProviders: protectedProcedure.query(({ ctx }) => {
     return ctx.paymentService.getPaymentProviders()
   }),
+  createStripeCheckoutSession: protectedProcedure
+    .input(
+      z.object({
+        productId: z.string().uuid(),
+        stripePublicKey: z.string(),
+        successRedirectUrl: z.string().url(),
+        cancelRedirectUrl: z.string().url(),
+      })
+    )
+    .mutation(({ input, ctx }) => {
+      return ctx.paymentService.createStripeCheckoutSessionForProductId(
+        input.productId,
+        input.stripePublicKey,
+        input.successRedirectUrl,
+        input.cancelRedirectUrl,
+        ctx.auth.userId
+      )
+    }),
 })
